refactor: update AvailableDevice imports to transport/types and remove l
Automated checkpoint by FlowPatch worker (iteration 1).
Card: #28 Implement the communication with electron
