+++ src/renderer/src/components/DeviceList.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable react-hooks/set-state-in-effect */
 import { useCallback, useEffect, useState } from 'react'
-import type { AvailableDevice } from '../tauri'
+import type { AvailableDevice } from '../transport/types'
 import { Key, ListBox, ListBoxItem, Selection } from 'react-aria-components'
 import { Bluetooth, RefreshCw } from 'lucide-react'
 import { TransportFactory } from '@/components/Modals/ConnectModal.tsx'
