+++ src/renderer/src/components/Modals/ConnectModal.tsx
@@ -1,7 +1,7 @@
 import { useMemo } from 'react'
 
 import type { RpcTransport } from '@zmkfirmware/zmk-studio-ts-client/transport/index'
-import type { AvailableDevice } from '../../tauri'
+import type { AvailableDevice } from '../../transport/types'
 import { ExternalLink } from '@/misc/ExternalLink.tsx'
 import { DeviceList } from '../DeviceList.tsx'
 import { SimpleDevicePicker } from '../SimpleDevicePicker.tsx'
